requirements.txt
flask
scikit-learn
pandas
joblib
requests
Pillow
